Add policy

+++ app/Domains/Admins/Actions/ReplyComments/DestroyReplyCommentAction.php
@@ -4,7 +4,7 @@ namespace Domains\Admins\Actions\ReplyComments;
 
 use App\Http\Requests\Admins\ReplyCommentRequest;
 use App\Models\Post;
-use Domains\Admins\DataTransferToObject\ReplyCommentData;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class DestroyReplyCommentAction
 {
@@ -16,7 +16,12 @@ class DestroyReplyCommentAction
 
         $replayComments = $comment->replyComments()->whereIn('id', $request->ids)->get();
 
+        $user = $request->user();
+
         foreach ($replayComments as $replayComment) {
+            if ($user->cannot('destroy', $replayComment)) {
+                throw new AuthorizationException();
+            }
             $replayComment->delete();
         }
     }
